changed TVC to VC becasue they are UIView Controllers not UITableView Controllers.
