+++ One Trick Companion/Controllers/Content/Maps/MapsViewController.swift
@@ -1,0 +1,28 @@
+//
+//  MapsViewController.swift
+//  One Trick Companion
+//
+//  Created by adam smith on 3/15/22.
+//
+
+import UIKit
+
+class MapsViewController: UIViewController {
+    
+//MARK: - Outlets
+    @IBOutlet weak var mapsTableView: UITableView!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        
+    }
+    
+
+
+    // MARK: - Navigation
+
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+
+    }
+}
