+++ One Trick Companion/Controllers/Leaderboard/LeaderboardTableViewCell.swift
@@ -1,0 +1,26 @@
+//
+//  LeaderboardTableViewCell.swift
+//  One Trick Companion
+//
+//  Created by adam smith on 3/15/22.
+//
+
+import UIKit
+
+class LeaderboardTableViewCell: UITableViewCell {
+
+//MARK: - Outlets
+    @IBOutlet weak var leaderboardRankTextLabel: UILabel!
+    @IBOutlet weak var agentNameTextlabel: UILabel!
+    @IBOutlet weak var rankedRatingTextLabel: UILabel!
+    @IBOutlet weak var rankTextLabel: UILabel!
+    
+    func updateViews() {
+        
+    }
+    
+    func fetchLeaderboard() {
+        
+    }
+    
+}
