+++ One Trick Companion/Controllers/Leaderboard/LeaderboardViewController.swift
@@ -1,0 +1,27 @@
+//
+//  LeaderboardViewController.swift
+//  One Trick Companion
+//
+//  Created by adam smith on 3/15/22.
+//
+
+import UIKit
+
+class LeaderboardViewController: UIViewController {
+    
+    //MARK: - Outlets
+    @IBOutlet weak var leaderboardTableView: UITableView!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+
+    // MARK: - Navigation
+
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+
+    }
+}
